+++ Assets/Scripts/Statki/Pancernik2.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 
 public class Pancernik2 : Ship
 {
+    int faza = 0;
     public Pancernik2() //konstruktor
     {
         name = "Pancernik";
@@ -16,20 +17,29 @@ public class Pancernik2 : Ship
 
     private void Awake()
     {
-        transform.position = new Vector3(PlayerPrefs.GetFloat("Pancernik2X"), 0.5f, PlayerPrefs.GetFloat("Pancernik2Z"));
-        fieldPosition = transform.position;
+        faza = PlayerPrefs.GetInt("Faza3");
+        if (faza == 0)
+        {
+            PlayerPrefs.DeleteKey("Pancernik2X");
+            PlayerPrefs.DeleteKey("Pancernik2Z");
+            faza = 1;
+            PlayerPrefs.SetInt("Faza3", faza);
+        }
+        else
+        {
+            transform.position = new Vector3(PlayerPrefs.GetFloat("Pancernik2X"), 0.5f, PlayerPrefs.GetFloat("Pancernik2Z"));
+            fieldPosition = transform.position;
+        }
     }
 
     private void OnApplicationQuit()
     {
-        PlayerPrefs.DeleteKey("Pancernik2X");
-        PlayerPrefs.DeleteKey("Pancernik2Z");
+        PlayerPrefs.DeleteAll();
     }
 
     private void Update()
     {
         PlayerPrefs.SetFloat("Pancernik2X", transform.position.x);
         PlayerPrefs.SetFloat("Pancernik2Z", transform.position.z);
-        //PlayerPrefs.Save();
     }
 }
